Fix video resume
Before video always started from the time argument in URL. Which means no session resume time.
Now video starts from where we left off. And it starts from the time argument in URL if loaded for the first time during the session or form was submitted again with time argument.