Closes issue#222: Add autoplay from Youtube
Cleaned up the code from pull request #231.
Then fixed some bugs:
(1) Fixed getNewVideoID function. Previously it just popped up all
the playList and left only the first element. And if this element
would be current video id then nextId will be set to null. Now I've
made a reverse for playlist array to get just the same next video from
YouTube as Youtube suggests itself and that is done with no further
changes in the code. I've also removed the while loop from
getNewVideoID function and that gives the code an opportunity
to check all the playlist for a next track during the session.
(2) Fixed the problem when tracks were added to videoMetadata["items"]
when they were already there. Now there is a clause that checks if
current track has already been played and if so just doesn't add
a duplicate.